+++ examples/agnostic/manifest.ts
@@ -2,7 +2,7 @@ import { Manifest } from "./deps.ts";
 import erc20 from "./abis/erc20.ts";
 import { transferHandler } from "./handlers/transfer.ts";
 
-const manifest = new Manifest();
+const manifest = new Manifest("agnostic-events");
 
 manifest
   .addChain("avalanche")