+++ src/arkiver/types.ts
@@ -48,6 +48,7 @@ export interface ArkiveManifest {
   >;
   // deno-lint-ignore no-explicit-any
   entities: { model: mongoose.Model<any>; list: boolean }[];
+  name: string;
 }
 
 export type DataSource = {